This completes the current work on importing bank and cash demo data
The import script brings in all the bank and cash demo date
The templates have been altered so that all the data shows
There are still some empty columns in the bank and cash views
i.e. where there are placeholders, but no demo data is given
There is still some data in the spreadsheet that is not shown in
the views
i.e. vat, and derived data such as totals and net/gross values